refactor(backgroundService): moved background services from features
