+++ src/Services/Parser/Parser.Core.Application/Parser.Core.Application.csproj
@@ -8,7 +8,8 @@
 
   <ItemGroup>
     <PackageReference Include="AutoMapper" Version="11.0.1" />
-    <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="11.0.0" />
+    <PackageReference Include="MediatR" Version="10.0.1" />
+    <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="10.0.1" />
     <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="6.0.0" />
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="6.0.1" />
   </ItemGroup>
@@ -17,6 +18,10 @@
     <ProjectReference Include="..\..\..\BuildingBlocks\EventBus\EventBus.RabbitMq\EventBus.RabbitMq.csproj" />
     <ProjectReference Include="..\Parser.Infrastructure.DataAccess\Parser.Infrastructure.DataAccess.csproj" />
     <ProjectReference Include="..\Parser.Infrastructure.ParserService\Parser.Infrastructure.HtmlAgilityPackService.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Features\" />
   </ItemGroup>
 
 </Project>
